--- a/dirty_cat/minhash_encoder.py
+++ b/dirty_cat/minhash_encoder.py
@@ -24,10 +24,11 @@ from sklearn.cluster import KMeans
 from sklearn.feature_extraction.text import CountVectorizer, HashingVectorizer
 from sklearn.neighbors import NearestNeighbors
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.utils import check_random_state, murmurhash3_32 
+from sklearn.utils import check_random_state, murmurhash3_32
 
 from .fast_hash import ngram_min_hash
 from .utils import LRUDict
+
 
 class MinHashEncoder(BaseEstimator, TransformerMixin):
     """
@@ -48,21 +49,29 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
         might have some concern with its entropy.
     minmax_hash : bool, default=False
         if True, return min hash and max hash concatenated.
+        handle_missing : 'error' or '' (default)
+        Whether to raise an error or impute with blank string '' if missing
+        values (NaN) are present during fit (default is to impute).
+        When this parameter is set to '', and a missing value is encountered
+        during fit_transform, the resulting encoded columns for this feature
+        will be all zeros.
     References
     ----------
     For a detailed description of the method, see
     `Encoding high-cardinality string categorical variables
     <https://hal.inria.fr/hal-02171256v4>`_ by Cerda, Varoquaux (2019).
-    
+
     """
 
     def __init__(self, n_components, ngram_range=(2, 4),
-                 hashing='fast', minmax_hash=False):
+                 hashing='fast', minmax_hash=False,
+                 handle_missing=''):
         self.ngram_range = ngram_range
         self.n_components = n_components
         self.hashing = hashing
         self.minmax_hash = minmax_hash
         self.count = 0
+        self.handle_missing = handle_missing
 
     def get_unique_ngrams(self, string, ngram_range):
         """ Return the set of unique n-grams of a string.
@@ -113,11 +122,10 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
             min_hashes = np.minimum(min_hashes, hash_array)
         return min_hashes/(2**32-1)
 
-    def get_hash(self, string):
+    def get_fast_hash(self, string):
         """
-        Encode a string with murmur or fast hashing function.
-        fast hashing supports both min_hash and minmax_hash encoding,
-        whereas murmur only supports min_hash encoding.
+        Encode a string with fast hashing function.
+        fast hashing supports both min_hash and minmax_hash encoding.
         Parameters
         ----------
         string : str
@@ -127,27 +135,30 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
         array, shape (n_components, )
             The encoded string, using specified encoding scheme.
         """
-        if self.hashing == 'fast':
-            if self.minmax_hash:
-                assert self.n_components % 2 == 0,\
-                       "n_components should be even when minmax_hash=True"
-                return np.concatenate([ngram_min_hash(string, self.ngram_range,
-                                                      seed, return_minmax=True)
-                                      for seed in range(self.n_components//2)])
-            else:
-                return np.array([ngram_min_hash(string, self.ngram_range, seed)
-                                for seed in range(self.n_components)])
+        if self.minmax_hash:
+            return np.concatenate([ngram_min_hash(string, self.ngram_range,
+                                   seed, return_minmax=True)
+                                   for seed in range(self.n_components//2)])
+        else:
+            return np.array([ngram_min_hash(string, self.ngram_range, seed)
+                            for seed in range(self.n_components)])
 
-        elif self.hashing == 'murmur':
-            assert not(self.minmax_hash),\
-                   "minmax_hash not implemented with murmur"
-            return self.minhash(
+    def get_murmur_hash(self, string):
+        """
+        Encode a string with murmur hashing function.
+        murmur only supports min_hash encoding.
+        Parameters
+        ----------
+        string : str
+            The string to encode.
+        Returns
+        -------
+        array, shape (n_components, )
+            The encoded string, using specified encoding scheme.
+        """
+        return self.minhash(
                     string, n_components=self.n_components,
                     ngram_range=self.ngram_range)
-        else:
-            raise ValueError("hashing function must be 'fast' or"
-                             "'murmur', got '{}'"
-                             "".format(self.hashing))
 
     def fit(self, X, y=None):
         """
@@ -157,11 +168,11 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
         ----------
         X : array-like, shape (n_samples, ) or (n_samples, 1)
             The string data to encode.
-        
+
         Returns
         -------
         self
-            The fitted MinHashEncoder instance. 
+            The fitted MinHashEncoder instance.
         """
         self.hash_dict = LRUDict(capacity=2**10)
         return self
@@ -184,14 +195,48 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
             X = X[:, 0]
         # Check if first item has str or np.str_ type
         assert isinstance(X[0], str), "ERROR: Input data is not string."
+        if self.minmax_hash:
+            assert self.n_components % 2 == 0,\
+                    "n_components should be even when minmax_hash=True"
+        if self.hashing == 'murmur':
+            assert not(self.minmax_hash),\
+                   "minmax_hash not implemented with murmur"
+        if self.handle_missing not in ['error', '']:
+            template = ("handle_missing should be either 'error' or "
+                        "'', got %s")
+            raise ValueError(template % self.handle_missing)
         X_out = np.zeros((len(X), self.n_components))
+        X = X.ravel()
 
         # TODO Parallel run here
-        for i, x in enumerate(X):
-            if x not in self.hash_dict:
-                self.hash_dict[x] = self.get_hash(x)
+        nan_idx = []
+
+        if self.hashing == 'fast':
+            for i, x in enumerate(X):
+                if isinstance(x, float):
+                    nan_idx.append(i)
+                elif x not in self.hash_dict:
+                    self.hash_dict[x] = self.get_fast_hash(x)
+
+        elif self.hashing == 'murmur':
+            for i, x in enumerate(X):
+                if isinstance(x, float):
+                    nan_idx.append(i)
+                elif x not in self.hash_dict:
+                    self.hash_dict[x] = self.get_murmur_hash(x)
+
+        else:
+            raise ValueError("hashing function must be 'fast' or"
+                             "'murmur', got '{}'"
+                             "".format(self.hashing))
+
+        if self.handle_missing == 'error' and nan_idx:
+            msg = ("Found missing values in input data; set "
+                   "handle_missing='' to encode with missing values")
+            raise ValueError(msg)
 
         for i, x in enumerate(X):
-            X_out[i, :] = self.hash_dict[x]
-            
+            if i not in nan_idx:
+                X_out[i, :] = self.hash_dict[x]
+
         return X_out