References Issue #124 Add support for missing values in the encoders

Changes: Added missing values support for minhash encoder