@amy12xx 2020-07-31
Hi @GaelVaroquaux, here's an attempt at it. there's probably room for improvement. I used the profile_encoder() and main() in test to check the perf.

BEFORE:
------------------------------------------------------------
start test
test passed
time profile_encoder(MinHashEncoder, hashing=fast)
0.6618 seconds >> 0.6987 seconds >> 0.7474 seconds
time profile_encoder(MinHashEncoder, hashing=fast)
0.6814 seconds >> 0.9635 seconds
time profile_encoder(MinHashEncoder, hashing=fast)
0.6376 seconds >> 0.7556 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.5987 seconds >> 0.5887 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.5654 seconds >> 0.7014 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.5932 seconds >> 0.6894 seconds
time profile_encoder(MinHashEncoder, hashing=murmur)
1.427 seconds >> 1.954
Done


AFTER:
------------------------------------------------------------
start test
test passed
time profile_encoder(MinHashEncoder, hashing=fast)
0.723 seconds
time profile_encoder(MinHashEncoder, hashing=fast)
0.691 seconds
time profile_encoder(MinHashEncoder, hashing=fast)
0.709 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.643 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.607 seconds
time profile_encoder(MinHashEncoder, hashing=fast) with minmax
0.6 seconds
time profile_encoder(MinHashEncoder, hashing=murmur)
1.506 seconds
Done